refactor: name shared feature helpers utils.ts and constants.ts
Domain-root helpers get one utils.ts / constants.ts per feature rather
than a per-component prefixed name.
